feat: When verifying user's email, set jwt cookie received from BE
